hcl colorpanel, cursor fix
