+++ ceammc/gui/plugins/hcs.colorpanel.tcl
@@ -1,0 +1,8 @@
+
+namespace eval ::hcs::colorpanel:: {
+}
+
+proc ::hcs::colorpanel::open {objectid initialcolor} {
+    set color [tk_chooseColor -initialcolor $initialcolor]
+    pdsend "$objectid callback $color"
+}
